refactor #29 styled-components 이름 수정 / handleOnBlur -> handleBlur 수정
