+++ src/pages/create-post-page.jsx
@@ -42,7 +42,6 @@ const ButtonWrapper = styled.div`
   display: flex;
   justify-content: center;
   align-items: center;
-
   & > * {
     flex: 1;
   }
