+++ src/pages/send-message-page.jsx
@@ -1,5 +1,91 @@
+import { useState } from "react";
+import InputTextField from "../components/text-field/text-field";
+import TEXT_FIELD_TYPE from "../components/text-field/text-field-type";
+import Colors from "../components/color/colors";
+import ToggleButton from "../components/button/toggle-button";
+import styled from "styled-components";
+
+const SendContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  margin: 0 auto;
+`;
+
+const Wrapper = styled.div`
+  padding-top: 50px;
+  width: 720px;
+`;
+
+const SendTitle = styled.h2`
+  font-weight: 700;
+`;
+
+const SendSummary = styled.p`
+  font-weight: 400;
+  color: Colors.gray(500);
+`;
+
+const ToggleButtonWrapper = styled.div`
+  width: 100%;
+  display: flex;
+`;
+
 function SendMessagePage() {
-  return <h1>Send Message Page </h1>;
+  const [name, setName] = useState("");
+  const [nameError, setNameError] = useState("");
+
+  const handleChange = (e) => {
+    const value = e.target.value;
+    setName(value);
+
+    // 값 입력 중 에러 없애기
+    if (nameError) {
+      setNameError("");
+    }
+  };
+
+  const handleBlur = () => {
+    if (name.trim() === "") {
+      setNameError("값을 입력해 주세요");
+    } else if (name !== name.trim()) {
+      setNameError("공백 없이 입력해 주세요"); // 텍스트 앞 뒤 공백 에러 처리(임시)
+    }
+  };
+
+  return (
+    <SendContainer>
+      <Wrapper>
+        <SendTitle>From.</SendTitle>
+        <InputTextField // TextField로 변경
+          type={TEXT_FIELD_TYPE.input}
+          value={name}
+          onChange={handleChange}
+          onBlur={handleBlur}
+          placeholder="이름을 입력해 주세요"
+          error={nameError}
+        />
+      </Wrapper>
+      <Wrapper>
+        <SendTitle>프로필 이미지</SendTitle>
+        <SendSummary>프로필 이미지를 선택해 주세요!</SendSummary>
+      </Wrapper>
+      <Wrapper>
+        <SendTitle>상대와의 관계</SendTitle>
+      </Wrapper>
+      <Wrapper>
+        <SendTitle>내용을 입력해 주세요</SendTitle>
+      </Wrapper>
+      <Wrapper>
+        <SendTitle>폰트 선택</SendTitle>
+      </Wrapper>
+      <ToggleButtonWrapper>
+        <ToggleButton value="컬러" options={["컬러", "이미지"]} />
+      </ToggleButtonWrapper>
+    </SendContainer>
+  );
 }
 
 export default SendMessagePage;
